backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==4.2.9
djangorestframework==3.14.0
dnspython==2.4.2
pykerberos==1.2.4
pymongo==4.6.1
python-dotenv==1.0.0
python-snappy==0.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
